Patch 12.4 build

+++ Sources/Private/CoreAnimation/Animations/ShapeAnimation.swift
@@ -136,8 +136,8 @@ extension Trim {
     // so the total path length includes the maximum stroke
     let startStrokes = adjustedStrokeEnd.keyframes.map { $0.value.cgFloatValue }
     let endStrokes = adjustedStrokeEnd.keyframes.map { $0.value.cgFloatValue }
-    let minimumStrokeMultiplier = floor((startStrokes.min() ?? 0) / 100.0)
-    let maximumStrokeMultiplier = ceil((endStrokes.max() ?? 100) / 100.0)
+    let minimumStrokeMultiplier = Double(floor((startStrokes.min() ?? 0) / 100.0))
+    let maximumStrokeMultiplier = Double(ceil((endStrokes.max() ?? 100) / 100.0))
 
     if minimumStrokeMultiplier < 0 {
       // Core Animation doesn't support negative stroke offsets, so we have to
